rename: 구체적인 컴포넌트 이름으로 수정 - SideNavItem 컴포넌트 (#52)
