--- src/components/layout/sidebar/NavItem/index.ts
@@ -1,1 +1,0 @@
-export { default as NavItem } from './NavItem';
