+++ src/components/layout/sidebar/SideNavItem/index.ts
@@ -1,0 +1,1 @@
+export { default as SideNavItem } from './SideNavItem';
